Consolidate some list styling
