+++ packages/web/components/patterns/LibraryCards/LibraryCardStyles.tsx
@@ -1,3 +1,8 @@
+import dayjs from 'dayjs'
+import relativeTime from 'dayjs/plugin/relativeTime'
+
+dayjs.extend(relativeTime)
+
 export const MetaStyle = {
   width: '100%',
   color: '$thTextSubtle3',
@@ -63,4 +68,40 @@ export const AuthorInfoStyle = {
   fontSize: '13px',
   fontWeight: '400',
   fontFamily: '$display',
+}
+
+export const timeAgo = (date: string | undefined): string => {
+  if (!date) {
+    return ''
+  }
+  return dayjs(date).fromNow()
+}
+
+const shouldHideUrl = (url: string): boolean => {
+  try {
+    const origin = new URL(url).origin
+    const hideHosts = ['https://storage.googleapis.com', 'https://omnivore.app']
+    if (hideHosts.indexOf(origin) != -1) {
+      return true
+    }
+  } catch {
+    console.log('invalid url item', url)
+  }
+  return false
+}
+
+export const siteName = (
+  originalArticleUrl: string,
+  itemUrl: string
+): string => {
+  if (shouldHideUrl(originalArticleUrl)) {
+    return ''
+  }
+  try {
+    return new URL(originalArticleUrl).hostname.replace(/^www\./, '')
+  } catch {}
+  try {
+    return new URL(itemUrl).hostname.replace(/^www\./, '')
+  } catch {}
+  return ''
 }
