+++ packages/web/components/patterns/LibraryCards/LibraryGridCard.tsx
@@ -13,6 +13,8 @@ import {
   DescriptionStyle,
   MenuStyle,
   MetaStyle,
+  siteName,
+  timeAgo,
   TitleStyle,
 } from './LibraryCardStyles'
 
@@ -46,39 +48,6 @@ export function ProgressBar(props: ProgressBarProps): JSX.Element {
       />
     </Box>
   )
-}
-
-const timeAgo = (date: string | undefined): string => {
-  if (!date) {
-    return ''
-  }
-  return dayjs(date).fromNow()
-}
-
-const shouldHideUrl = (url: string): boolean => {
-  try {
-    const origin = new URL(url).origin
-    const hideHosts = ['https://storage.googleapis.com', 'https://omnivore.app']
-    if (hideHosts.indexOf(origin) != -1) {
-      return true
-    }
-  } catch {
-    console.log('invalid url item', url)
-  }
-  return false
-}
-
-const siteName = (originalArticleUrl: string, itemUrl: string): string => {
-  if (shouldHideUrl(originalArticleUrl)) {
-    return ''
-  }
-  try {
-    return new URL(originalArticleUrl).hostname.replace(/^www\./, '')
-  } catch {}
-  try {
-    return new URL(itemUrl).hostname.replace(/^www\./, '')
-  } catch {}
-  return ''
 }
 
 // Component
